Register geolocation serializer only if plone.restapi is installed
